application/requirements.txt
streamlit
pyspark==3.4.3
matplotlib
seaborn
snowflake-connector-python